requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic
email-validator